♻️ Refactor of the plaquette library
